+++ cfn_datadog/__init__.py
@@ -1,6 +1,6 @@
 from troposphere.validators import boolean, positive_integer
-from troposphere import cloudformation, depends_on_helper
-from troposphere.cloudformation import AWSHelperFn, AWSObject
+from troposphere import cloudformation, depends_on_helper, BaseAWSObject
+from troposphere.cloudformation import AWSHelperFn, AWSObject, AWSProperty
 import types, sys
 
 try:
@@ -12,7 +12,28 @@ except NameError:
     bytes = str
 
 
+class CustomProperty(AWSProperty):
+    """
+    Custom property to be able to use title in a property,
+    this is needed for the timeboards because they use the title property.
+    """
+    dictname = None
+
+    def __init__(self, title=None, **kwargs):
+        self.title = None
+        print("title")
+        print(title)
+
+        print("self.title")
+        print(self.title)
+
+        print("kwargs")
+        print(kwargs)
+        super().__init__(**kwargs)
+
+
 class DatadogBase(AWSObject):
+
     """
     Removed condition to not validate if based on custom:
     if type_name == 'AWS::CloudFormation::CustomResource' or \
@@ -101,4 +122,4 @@ from .metric_alert import MetricAlert, MetricAlertOptions, Thresholds
 from .service_check import ServiceCheck, ServiceCheckOptions
 from .composite import CompositeOptions, Composite
 from .event_alert import EventAlertOptions, EventAlert
-from .timeboard import TimeBoard
+from .timeboard import TimeBoard, Graph, TemplateVariable, Definition, Request